+++ tests/commands/test_upgrade.py
@@ -39,7 +39,7 @@ def test_command_release_tenderers_amendment(monkeypatch, caplog):
     assert len(caplog.records) == 1
     assert caplog.records[0].levelname == 'WARNING'
     assert caplog.records[0].message == (
-        'party differs in "supplier" role than in "tenderer" roles:\n'
+        'party in "supplier" role differs from party in ["tenderer"] roles:\n'
         '{"name": "Acme Inc.", "identifier": {"id": 1}, "additionalIdentifiers": [{"id": "a"}], "id": "3c9756cf8983b14066a034079aa7aae4"}\n'  # noqa: E501
         '{"id": "3c9756cf8983b14066a034079aa7aae4", "name": "Acme Inc.", "identifier": {"id": 1}}'
     )